move test_detail in test_stitching
